Rework usrmerge support and add profiling flags
The previous usrmerge support was fairly hacky, we ended up
with duplicate packages per pid, requiring us to filter them
again with a set-like map. Instead of that crap, just trim of
/usr from executables starting with /usr and check if we can
find it that way if we did not find it yet.

This commit also adds some flags for the profiling options so
we do not profile all the time, adds some headers and borders
to the output, and increases the file->pkg map size.