Get rid of unsafe again
The strconv one was actually safe-ish: The memory referenced
was not referenced again after the strconv call. Just using a
string however does not make it show up in the profile, so it's
pointless performance wise.

The second one actually looks unsafe: From the description it
would appear that the memory is reused again and again, and
given that we store the per-line strings in a struct, it should
not work at all. It did, however, work... That said, let's just
fix this, it's ugly to use unsafe.